更新SDK代码

+++ cpp/include/TANetwork.h
@@ -15,7 +15,7 @@ namespace TaSDK {
         TANetwork(void);
         ~TANetwork(void);
         
-        int post(const string &url, const string &appid, string &data, int32_t dataSize, string &strResponse);
+        int64_t post(const string &url, const string &appid, string &data, int32_t dataSize, string &strResponse);
         int64_t debug_post(const string &url, const string &appid, string &data, int32_t dataSize, string &strResponse, int32_t dryRun);
     };
 };
